requirements.txt
flask
python-telegram-bot
apscheduler
tradingview-scraper
